requirements.txt
# Core LLM
anthropic>=0.39.0

# RAG pipeline
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Agent orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# API layer
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0          # async HTTP client for web search
beautifulsoup4>=4.12.0 # HTML parsing for web search

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
